+++ src/styles/TodoListItem.module.css
@@ -14,7 +14,7 @@
     text-align: center;
     max-width: 900px;
     margin: 20px auto;
-    background-image: url('./img/todolist.png');
+    background-image: url('../img/todolist.png');
    
   }
   .MainHeader {
