+++ src/styles/TodoView.module.css
@@ -1,0 +1,73 @@
+:root {
+    --background-color: #cde9ff;
+    --secondary-color: #e4f3ff;
+    --hover-color: #0065b3;
+    --main-font: "Courier New", Courier, monospace;
+    --main-font-color: #00335ad3;
+    --transition: ease-in-out 125ms;
+    --border-radius: 4px;
+    --box-shadow: 0px 1px 1px;
+  }
+  
+  button {
+    background-color: var(--secondary-color);
+    font-family: var(--main-font);
+    border: 1px solid black;
+    border-color: var(--background-color);
+    border-radius: var(--border-radius);
+    color: var(--main-font-color);
+    box-shadow: var(--box-shadow);
+    padding: 15px;
+  }
+  
+  button:hover {
+    cursor: pointer;
+    background-color: var(--hover-color);
+    color: var(--background-color);
+    transition: var(--transition);
+  }
+  
+  body {
+    background-color: #cde9ff;
+    font-family: var(--main-font);
+    color: var(--main-font-color);
+  }
+  
+  .Container {
+    padding: 20px;
+    text-align: center;
+    max-width: 900px;
+    margin: 20px auto;
+    background-image: url('../img/todolist.png');
+  }
+  .MainHeader {
+    color: var(--main-font-color);
+    font-size: 44px;
+  }
+  
+  .Loading {
+    color: var(--main-font-color);
+    font-family: var(--main-font);
+  }
+  
+  .SortButton {
+    margin: 10px;
+  }
+  
+  @media (max-width: 976px) {
+    .FormInputField {
+      padding: 12px 0px;
+    }
+    .MainHeader {
+      font-size: 35px;
+    }
+    .AddTodoFormButton {
+      padding: 10px;
+    }
+  }
+  
+  @media (max-width: 600px) {
+    body {
+      font-size: 24px;
+    }
+  }
